requirements.txt
altair
annotated-types
anyio
attrs
bitarray
blinker
cachetools
catalogue
certifi
charset-normalizer
click
colorama
document
docx
exceptiongroup
gitdb
GitPython
h11
httpcore
httpx
idna
jamaibase
Jinja2
jsonschema
jsonschema-specifications
loguru
lxml
markdown-it-py
MarkupSafe
mdurl
narwhals
numpy
orjson
packaging
pandas
pdfreader
pillow
protobuf
pyarrow
pycryptodome
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyPDF2
python-dateutil
python-docx
python-dotenv
pytz
referencing
requests
rich
rpds-py
six
smmap
sniffio
streamlit
tenacity
toml
tornado
typing_extensions
tzdata
urllib3
watchdog
win32-setctime
